humansize: rework types entirely
Due to broken platforms, support a varying range of possible floating
point types, using _generic where available to avoid needing to actually
change anything, for the most part.  For better or worse, focus on
double usage where possible.

+++ test/ssb.c
@@ -5,13 +5,13 @@
 
 static void print_ssb_info(struct ssb *s)
 {
-	long double size, pos;
+	double size, pos;
 	char *size_pre, *pos_pre;
 
 	humansize_scale(s->pos, &humansize_iec, &pos, &pos_pre);
 	humansize_scale(s->size, &humansize_iec, &size, &size_pre);
 
-	fprintf(stderr, "\t\tgrow strategy %d: position %Lf%sB, size %Lf%sB\n", s->grow,
+	fprintf(stderr, "\t\tgrow strategy %d: position %f%sB, size %f%sB\n", s->grow,
 			pos, pos_pre, size, size_pre);
 }
